fix ecs_get_version() test

+++ test/core/src/Entity.c
@@ -2290,7 +2290,7 @@ void Entity_get_version_after_reuse(void) {
 
     ecs_entity_t e_new = ecs_new(world);
     uint32_t v1 = ecs_get_version(e_new);
-    test_assert(e & 0xFFFFFFFF == e_new & 0xFFFFFFFF);
+    test_assert((uint32_t)(e & 0xFFFFFFFF) == (uint32_t)(e_new & 0xFFFFFFFF));
     test_assert(v1 == v0 + 1);
 
     ecs_fini(world);
